Refactor: move out enterprise fork sync code
